requirements.txt
discord.py[voice]
python-dotenv
google-generativeai
gTTS